+++ sdk/webpubsub-socketio-extension/CHANGELOG.md
@@ -1,7 +1,15 @@
 
 # Changelog
 
-## [1.2.1-beta.1]
+## [2.0.0] - 2026-08-21
+
+### Breaking Changes
+
+- Require Node.js 22 or later.
+
+### Fixed
+
+- Prevent idle tunnel connections from reconnecting when using `@azure/web-pubsub-client` 1.0.4 or later [#1020](https://github.com/Azure/azure-webpubsub/issues/1020)
 
 ## [1.2.1] - 2026-04-22